refactor(dashboard): move shared client surface to crate::shared::client::* (#578 Phase C)
Move cross-app SPA building blocks out of dashboard/src/client/ into
dashboard/src/shared/client/* so each app owns only its own pages and
the eventual deletion of dashboard/src/client/ in Phase E becomes a
purely import-path concern. Files relocated:

  client/pages/not_found.rs          -> shared/client/pages/not_found.rs
  client/components/toast.rs         -> shared/client/components/toast.rs
  client/components/status_badge.rs  -> shared/client/components/status_badge.rs
  client/state.rs                    -> shared/client/state.rs
  client/ws.rs                       -> shared/client/ws.rs

dashboard/src/client/router.rs is intentionally untouched - Phase D
will rewrite it to aggregate per-app url_patterns(). Phases A and B
landed in #580 and #581.

Refs #578
