+++ dashboard/src/apps/deployments/client/components/cluster_health.rs
@@ -14,7 +14,7 @@ use reinhardt::pages::page;
 use crate::shared::ws_messages::ClusterHealthPayload;
 
 #[cfg(wasm)]
-use crate::client::components::toast::html_escape;
+use crate::shared::client::components::toast::html_escape;
 
 /// DOM id of the cluster health container.
 #[cfg(wasm)]
