--- dashboard/src/client/components.rs
@@ -1,4 +1,0 @@
-//! Client-side UI components for the dashboard.
-
-pub mod status_badge;
-pub mod toast;
