--- dashboard/src/client/pages.rs
@@ -1,5 +1,0 @@
-//! Cross-app page components.
-
-pub mod not_found;
-
-pub use not_found::not_found_page;
