+++ dashboard/src/shared/client/components.rs
@@ -1,0 +1,4 @@
+//! Cross-app client-side UI components for the dashboard SPA.
+
+pub mod status_badge;
+pub mod toast;
