+++ dashboard/src/shared/client/components/status_badge.rs
