+++ dashboard/src/shared/client/components/toast.rs
