+++ dashboard/src/shared/client/pages.rs
@@ -1,0 +1,3 @@
+//! Cross-app page components for the dashboard SPA.
+
+pub mod not_found;
