+++ dashboard/src/shared/client/pages/not_found.rs
