+++ dashboard/src/shared/client/state.rs
