+++ dashboard/src/shared/client/ws.rs
